+++ src/charactersData/index.ts
@@ -43,6 +43,7 @@ import Yoimiya from './characters/Yoimiya';
 import Traveler from './characters/Traveler';
 import Thoma from './characters/Thoma';
 import Gorou from './characters/Gorou';
+import AratakiItto from './characters/AratakiItto';
 
 // TODO: не забыть добавить путешественника
 const CHARACTERS: ICharacter[] = [
@@ -88,7 +89,8 @@ const CHARACTERS: ICharacter[] = [
   Yoimiya,
   Traveler,
   Thoma,
-  Gorou
+  Gorou,
+  AratakiItto
 ]
 
 export default CHARACTERS