+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     * Hu Tao
     * Xiao
     * Eula
+    * Tartaglia
+    * Raiden
 
 #### 1.0.3 November-2021