+++ CHANGELOG.md
@@ -3,6 +3,12 @@
   * Added possible teams section
   * Added possible teams for:
     * Albedo
+    * Beidou
+    * Hu Tao
+    * Xiao
+    * Eula
+    * Tartaglia
+    * Raiden
 
 #### 1.0.3 November-2021