+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * Tartaglia
     * Raiden
   * Added Itto and Gorou
+  * Added Golden Wolflord material
 
 #### 1.0.3 November-2021