+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * Tartaglia
     * Raiden
     * Ayaka
+    * Ganyu
   * Added Itto and Gorou
   * Added Golden Wolflord material
   * Fixed talent level-up materials for Traveler