+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * Raiden
     * Ayaka
     * Ganyu
+    * Yanfei
   * Added Itto and Gorou
   * Added Golden Wolflord material
   * Fixed talent level-up materials for Traveler