+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     * Ayaka
     * Ganyu
     * Yanfei
+    * Klee
   * Added Itto and Gorou
   * Added Golden Wolflord material
   * Fixed talent level-up materials for Traveler