+++ CHANGELOG.md
@@ -14,6 +14,7 @@
     * Yanfei
     * Klee
     * Itto
+    * Diluc
   * Added Itto and Gorou
   * Added Golden Wolflord material
   * Fixed talent level-up materials for Traveler