+++ CHANGELOG.md
@@ -9,6 +9,13 @@
     * Eula
     * Tartaglia
     * Raiden
+    * Ayaka
+    * Ganyu
+    * Yanfei
+    * Klee
+    * Itto
+    * Gorou
+    * Diluc
   * Added Itto and Gorou
   * Added Golden Wolflord material
   * Fixed talent level-up materials for Traveler