+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#### 1.1.5 January-2022
+* Changes
+  * Added possible teams for:
+    * Yoimiya
+
 #### 1.1.4 December-2021
 * Changes
   * Added possible teams section